Moving the complex test for 93e2f24c62 into a separate file and allowing it to run in parallel with other tests
