+++ core/src/test/java/org/jclouds/reflect/Reflection2Test.java
@@ -19,17 +19,12 @@
 package org.jclouds.reflect;
 
 import static com.google.common.base.Functions.toStringFunction;
-import static com.google.common.base.Throwables.propagate;
-import static java.lang.String.format;
 import static org.jclouds.reflect.Reflection2.constructors;
 import static org.jclouds.reflect.Reflection2.method;
 import static org.jclouds.reflect.Reflection2.methods;
 import static org.jclouds.reflect.Reflection2.typeToken;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.fail;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.SortedSet;
@@ -37,9 +32,6 @@ import java.util.SortedSet;
 import org.testng.annotations.Test;
 
 import com.google.common.base.Function;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.reflect.Invokable;
@@ -121,97 +113,9 @@ public class Reflection2Test {
                   .addAll(setMethods).build());
    }
 
-   public void testOverriddenMethodWithNarrowedReturnType() throws NoSuchMethodException {
-       LoadingCache<TypeToken<?>, Set<Invokable<?, ?>>> methodsForTypeTokenBackup =
-           getStaticField(Reflection2.class, "methodsForTypeToken");
-       // expecting two methods: the declared "method" and the bridge version with return type Set
-       final Method[] methods = ChildOverridesAndNarrowsMethod.class.getDeclaredMethods();
-       try {
-           /*
-            * Force Reflection2.methodsForTypeToken to reflect the fact that the declared methods
-            * of a class are not returned in any particular order.
-            */
-           setStaticField(Reflection2.class, "methodsForTypeToken", CacheBuilder.newBuilder().build(
-                   new CacheLoader<TypeToken<?>, Set<Invokable<?, ?>>>() {
-                          @Override
-                           public Set<Invokable<?, ?>> load(TypeToken<?> key) throws Exception {
-                              if (!key.equals(TypeToken.of(ChildOverridesAndNarrowsMethod.class))) {
-                                  fail(format("expected only key %s to be requested, but was %s", 
-                                          TypeToken.of(ChildOverridesAndNarrowsMethod.class), key));
-                              }
-                              return ImmutableSet.<Invokable<?, ?>>of(Invokable.from(methods[0]), Invokable.from(methods[1]));
-                           }
-                   }));
-           // getMethod returns the method with the *narrowest* return type if one exists
-           assertEquals(Reflection2.method(ChildOverridesAndNarrowsMethod.class, "method"), 
-                   Invokable.from(ChildOverridesAndNarrowsMethod.class.getMethod("method")));
-
-           // now the opposite order
-           Reflection2Test.<LoadingCache<?, ?>>
-               getStaticField(Reflection2.class, "methodForParams").invalidateAll();
-           setStaticField(Reflection2.class, "methodsForTypeToken", CacheBuilder.newBuilder().build(
-                   new CacheLoader<TypeToken<?>, Set<Invokable<?, ?>>>() {
-                          @Override
-                           public Set<Invokable<?, ?>> load(TypeToken<?> key) throws Exception {
-                              if (!key.equals(TypeToken.of(ChildOverridesAndNarrowsMethod.class))) {
-                                  fail(format("expected only key %s to be requested, but was %s", 
-                                          TypeToken.of(ChildOverridesAndNarrowsMethod.class), key));
-                              }
-                              return ImmutableSet.<Invokable<?, ?>>of(Invokable.from(methods[1]), Invokable.from(methods[0]));
-                           }
-                   }));
-           // getMethod returns the method with the *narrowest* return type if one exists
-           assertEquals(Reflection2.method(ChildOverridesAndNarrowsMethod.class, "method"), 
-                   Invokable.from(ChildOverridesAndNarrowsMethod.class.getMethod("method")));
-       } finally {
-           setStaticField(Reflection2.class, "methodsForTypeToken", methodsForTypeTokenBackup);
-       }
-   }
-
-   @SuppressWarnings("unchecked")
-   private static <T> T getStaticField(Class<?> declaringClass, String fieldName) {
-       try {
-            Field field = declaringClass.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            // static field
-            return (T) field.get(null);
-        } catch (NoSuchFieldException exception) {
-            throw propagate(exception);
-        } catch (IllegalAccessException exception) {
-            throw propagate(exception);
-        }
-   }
-
-   private static void setStaticField(Class<?> declaringClass, String fieldName, Object value) {
-       try {
-            Field field = declaringClass.getDeclaredField(fieldName);
-            field.setAccessible(true);
-            // static field
-            field.set(null, value);
-        } catch (NoSuchFieldException exception) {
-            throw propagate(exception);
-        } catch (IllegalAccessException exception) {
-            throw propagate(exception);
-        }
-   }
-
    static final Function<Invokable<?, ?>, String> invokableToName = new Function<Invokable<?, ?>, String>() {
       public String apply(Invokable<?, ?> input) {
          return input.getName();
       }
    };
-
-   private static class ParentWithMethod {
-       @SuppressWarnings("unused")
-       public Set<Object> method() {
-           return null;
-       }
-   }
-
-   private static class ChildOverridesAndNarrowsMethod extends ParentWithMethod {
-       @Override
-       public SortedSet<Object> method() {
-           return null;
-       }
-   }
 }
